requirements.txt
accelerate==0.16.0
datasets==2.9.0
evaluate==0.4.0
rouge-score==0.1.2
fastapi==0.92.0
sqlparse==0.4.3
huggingface_hub==0.12.0
open_clip_torch==2.14.0
nltk==3.8.1
numpy==1.22.4
pandas==1.3.4
Pillow==9.4.0
pydantic==1.8.2
requests==2.28.1
# --extra-index-url https://download.pytorch.org/whl/cu117
# torch==1.13.0+cu117
# torchvision==0.14.0+cu117
# torchaudio==0.13.0+cu117
torch==1.13.0
torchvision==0.14.0
torchaudio==0.13.0
git+https://github.com/huggingface/transformers.git
uvicorn==0.20.0
